Scraper to database test case works... kind of
Need to improve scraper parsing. The scraper will import data to some
models, but it needs to be more precise in its parsing to provide more
detailed information. Also, scraping and importation is not flawless:
there are some instances that models.location.address_name is stored as
an entire incident.